requirements.txt
streamlit
python-dotenv
pydantic
# AgnoAI SDK and related tools
agno
openai
fastapi
starlette
scrapegraph-py
